flask/requirements.txt
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
certifi==2025.1.31
charset-normalizer==3.4.1
click @ file:///C:/b/abs_f9ihnt72pu/croot/click_1698129847492/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
Flask @ file:///C:/b/abs_a6m37uz578/croot/flask_1737454335662/work
idna==3.10
itsdangerous @ file:///C:/b/abs_c4vwgdr5yn/croot/itsdangerous_1716533399914/work
Jinja2 @ file:///C:/b/abs_10zs3i4lof/croot/jinja2_1737760123028/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
numpy==1.23.5
pillow==11.1.0
requests==2.32.3
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
